// frontend/src/openapi/models/EvalRunProgress.ts
export interface EvalRunProgress {
    /**
     *
     * @type {PhaseEnum}
     * @memberof EvalRunProgress
     */
    phase: PhaseEnum;
    /**
     *
     * @type {number}
     * @memberof EvalRunProgress
     */
    prepared: number;
    /**
     *
     * @type {number}
     * @memberof EvalRunProgress
     */
    total: number;
    /**
     *
     * @type {number}
     * @memberof EvalRunProgress
     */
    errors: number;
    /**
     *
     * @type {number}
     * @memberof EvalRunProgress
     */
    scored: number;
    /**
     *
     * @type {number}
     * @memberof EvalRunProgress
     */
    scoreTotal: number;
    /**
     *
     * @type {Array<EvalRunVariantProgress>}
     * @memberof EvalRunProgress
     */
    variants: Array<EvalRunVariantProgress>;
    /**
     *
     * @type {Array<EvalRunEvaluatorStat>}
     * @memberof EvalRunProgress
     */
    evaluatorStats: Array<EvalRunEvaluatorStat>;
}